+++ server/src/main/java/server/database/TagRepository.java
@@ -18,4 +18,4 @@ package server.database;
 import commons.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TagRepository extends JpaRepository<Tag, Long> {}
+public interface TagRepository extends JpaRepository<Tag, Integer> {}
